+++ src/demo_tezos_events/handlers/on_wrap_event.py
@@ -1,10 +1,10 @@
 from demo_tezos_events import models as models
-from demo_tezos_events.types.events_contract.tezos_events.move import MovePayload
+from demo_tezos_events.types.events_contract.tezos_events.wrap import WrapPayload
 from dipdup.context import HandlerContext
 from dipdup.models.tezos import TezosEvent
 
 
-async def on_move_event(
+async def on_wrap_event(
     ctx: HandlerContext,
-    event: TezosEvent[MovePayload],
+    event: TezosEvent[WrapPayload],
 ) -> None: ...